+++ include/tmv/TMV_ProdMV.h
@@ -134,8 +134,8 @@ namespace tmv {
         Vector<typename V3::value_type> v3i = v3;
         Vector<typename V3::value_type> v3c = v3;
         if (!add) v3c.setZero();
-        for(int i=0;i<v3.size();++i) {
-            for(int j=0;j<v2.size();++j) {
+        for(size_t i=0;i<v3.size();++i) {
+            for(size_t j=0;j<v2.size();++j) {
                 v3c.ref(i) += T(x) * m1.cref(i,j) * v2.cref(j);
             }
         }
@@ -166,8 +166,8 @@ namespace tmv {
         Vector<typename V3::value_type> v3i = v3;
         Vector<typename V3::value_type> v3c = v3;
         if (!add) v3c.setZero();
-        for(int i=0;i<v1.size();++i) {
-            for(int j=0;j<v3.size();++j) {
+        for(size_t i=0;i<v1.size();++i) {
+            for(size_t j=0;j<v3.size();++j) {
                 v3c.ref(j) += T(x) * v1.cref(i) *  m2.cref(i,j);
             }
         }
@@ -198,8 +198,8 @@ namespace tmv {
         Vector<typename V1::value_type> v1i = v1;
         Vector<typename V1::value_type> v3 = v1;
         v3.setZero();
-        for(int i=0;i<v1.size();++i) {
-            for(int j=0;j<v1.size();++j) {
+        for(size_t i=0;i<v1.size();++i) {
+            for(size_t j=0;j<v1.size();++j) {
                 v3.ref(j) += T(x) * v1.cref(i) *  m2.cref(i,j);
             }
         }
